tst_ShellWithPin: Replace loader with temporary object
The reason for this is best discussed in [1], but as a summary:

There can only be one instance of SurfaceManager - and a few other
objects - at a time. If the tests run fast enough, it is possible for
the Loader which loaded the Shell to release the old Shell (which holds
a SurfaceManager) and then load the new shell before the old one is
finished destructing. This leads to an assertion failure in
SurfaceManager::SurfaceManager and the tests closing.

Fix this by using QtTest.TestCase.createTemporaryObject, which is made
specifically for creating objects that are deleted at the end of a test
run.

In order to do this, I had to keep the Shell from automatically loading at
the start of the TestCase when run as 'make tryShellWithPin'... so now
we get a button to load the Shell.

[1] https://github.com/ubports/unity8/pull/355#discussion_r556968153